+++ app/controllers/nominations/choose_how_to_continue_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Nominations::ChooseHowToContinueController < ApplicationController
+  include NominationEmailTokenConsumer
+
   before_action :check_token_status, only: :new
 
   def new
@@ -30,28 +32,6 @@ class Nominations::ChooseHowToContinueController < ApplicationController
 
 private
 
-  def check_token_status
-    if nomination_email.nil?
-      redirect_to link_invalid_nominate_induction_coordinator_path
-    elsif nomination_email.expired?
-      redirect_to link_expired_nominate_induction_coordinator_path(school_id: nomination_email.school_id)
-    elsif nomination_email.school.registered?
-      redirect_to already_nominated_request_nomination_invite_path
-    end
-  end
-
-  def token
-    @token ||= params[:token]
-  end
-
-  def nomination_email
-    @nomination_email ||= NominationEmail.find_by(token: token)
-  end
-
-  def school
-    nomination_email&.school
-  end
-
   def record_opt_out_state_and_redirect!
     opt_out = @how_to_continue_form.opt_out?
 
@@ -73,12 +53,6 @@ private
     form_params = params.require(:nominate_how_to_continue_form).permit(:how_to_continue, :token)
     @token = form_params[:token]
     form_params
-  end
-
-  def record_nomination_email_opened
-    NominationEmail
-      .where(token: nomination_email.token, opened_at: nil)
-      .update_all(opened_at: Time.zone.now)
   end
 
   def cohort
