+++ spec/cypress/support/step_definitions/common-interaction.js
@@ -11,6 +11,11 @@ const inputs = {
   "search box": "input[name=query]",
   "replace induction tutor": "input[value=replace].govuk-radios__input",
   "update induction tutor": "input[value=update].govuk-radios__input",
+  "nominate induction tutor radio button": "input[value=yes].govuk-radios__input",
+  "opt out of updates radio button": "input[value=no].govuk-radios__input",
+  "nominate someone for updates radio button": "input[value=i_dont_know].govuk-radios__input",
+  "design and deliver our own programme radio button": "input[value=design_our_own].govuk-radios__input",
+  "no early career teachers radio button": "input[value=no_early_career_teachers].govuk-radios__input",
 };
 
 const buttons = {
